Updating Travis CI node.js versions
It will now run the tests in the latest node.js v0.10.x (the previous behavior) AND v0.12.x, AND the current stable release, currently v4.0.0 (v1-3 was used by iojs, so they bumped to v4 when they merged it back into node.js. So the entire test suite will now be run three times and have to pass all three to get a "passing" build.