List breaking changes for v1
I suppose the STT createRecognizeStream() is not technically a breaking change since we decided to keep the old methods in place for now. But it will be in the near future, so I think it still makes sense to include here.

Document Conversion's API reduction was also intended to be a v1 breaking change, but looks like it leaked into master early. I'm still calling it out here.