[DCS] Makes default version_date a static value
We decided to make the default version date a static value for the
experimental service. This also allows us to remove the dependency on
moment.

This commit also changes how the version date is expected to be passed
in, moving it to a top-level option to be consistent with the Visual
Recognition v2-beta API.