eslint & related fixes
The project previously had jshint.. but it was misconfigured so that it didn't actually check any of the source.
So.. enabling it reported 200+ errors.

I installed eslint partially for the --fix option because it autofixes half of the issues (try npm run autofix ;)
After that, I fixed the remaining issues, and swapped jshint out for eslint.