docs: fix dead links in README
Noticed a couple of dead links in the README doc. One is a typo
(AUTHENTICATION.md should be Authentication.md) so I fixed that.

The other pointed at a doc on VCAP_SERVICES that doesn't exist
any more, so I changed it for the only description I could find
for VCAP_SERVICES on the public cloud docs site.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>